Fixed error_handler which now terminates script with exit() and exitstatus
This ensures operations are not reported as successful when there are Errors or Warnings (typically caused by changes in PHP core which cause breaking changes in SRDB). By adding exit() to the error_handler callback, wrapper scripts which call `srdb.cli.php` from the command line will properly detect when an error occurs to stop operations.